requirements.txt
# Updated requirements.txt
# Core packages
sqlalchemy==1.4.50
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.0
seaborn==0.12.2
scikit-learn==1.3.2
scipy==1.11.4

# RedShift specific
redshift-connector==2.0.914
sqlalchemy-redshift==0.8.14

# Utilities
python-dotenv==1.0.0
openpyxl==3.1.2
xlsxwriter==3.1.9
pyyaml==6.0.3
pydantic==2.5.0

# Optional for development
jupyter==1.0.0
ipykernel==6.25.2

pyarrow>=12.0.0
-e .